requirements-dev.txt
black>=22.1
django-stubs>=1.9
flake8>=4.0.1
mypy>=0.931
pytest-cov>=3.0
pytest-django>=4.5
